Persist data
"In this lesson, you added the ability to save and load the app’s data. This lets the data persist across multiple runs. Whenever the app launches, it loads the existing data. When the data is modified, the app saves it. The app can terminate safely without losing any data."